Refactor/khanh huy/view admin

+++ Techboys/database/migrations/2025_01_13_012955_create_bill_details_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('product_id')->nullable();
             $table->foreignId('variant_id')->nullable();
             $table->integer('quantity');
-            $table->decimal('price', 10, 2);
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
