+++ Techboys/database/migrations/2025_01_13_013721_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('img')->nullable();
             $table->string('slug');
-            $table->decimal('base_price', 20, 2);
+            $table->bigInteger('base_price');
             $table->integer('base_stock');
             $table->tinyInteger('is_featured');
             $table->decimal('rate_average', 10, 2);
